!fixup SystemUI: Fix showing duplicated charging status
* !fixup SystemUI: Fix showing duplicated charging status

it was possible that plugging a device in, then turning on the screen would show the charging status twice. To fix, don't append the charging status to if there is a transient notification.

* Don't shift keyguard when no fingerprint is added and hide FOD icon on strong authentication

Squashed from
Change-Id: Iccf469fa4fed79ff0edde014ba165eef6ee024e1
Change-Id: Ie227aee88fea4945a7618fe606c1094ae812ebe0
By Francesco "Dil3mm4" Manzo of Descendant

This commit disable shifting of keyguard when there is no fingerprint or if all
fingerprint is removed.
Tests:
    -removed enrolled fp, ensured keyguardsecuritycontainer wasn't shifed
    -removed enrolled fp, rebooted, ensured keyguardsecuritycontainer (on SIM PIN) wasn't shifted
This also hides the FOD icon when strong authentication is required. i.e - coming from cold boot or reboot

Co-authored-by: Scott Warner <Tortel1210@gmail.com>
Co-authored-by: Dil3mm4 <dil3mm4.dev@gmail.com>
Signed-off-by: rajkale99 <kaleraj.rk@gmail.com>